refactor: move accessibility/, helpers/, translations/ into core/
Moves accessibility utils to core/accessibility/, helper files to
core/util/ and core/widgets/, and translations extension to
core/translation/.
