+++ lib/features/document_details/view/widgets/document_share_button.dart
@@ -11,9 +11,9 @@ import 'package:paperless_mobile/features/document_details/cubit/document_detail
 import 'package:paperless_mobile/features/document_details/view/dialogs/select_file_type_dialog.dart';
 import 'package:paperless_mobile/core/model/file_download_type.dart';
 import 'package:paperless_mobile/generated/l10n/app_localizations.dart';
-import 'package:paperless_mobile/helpers/connectivity_aware_action_wrapper.dart';
-import 'package:paperless_mobile/helpers/message_helpers.dart';
-import 'package:paperless_mobile/helpers/permission_helpers.dart';
+import 'package:paperless_mobile/core/widgets/connectivity_aware_action_wrapper.dart';
+import 'package:paperless_mobile/core/util/message_helpers.dart';
+import 'package:paperless_mobile/core/util/permission_helpers.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:provider/provider.dart';
