+++ lib/features/document_scan/view/scanner_page.dart
@@ -24,9 +24,9 @@ import 'package:paperless_mobile/features/document_search/view/sliver_search_bar
 import 'package:paperless_mobile/features/document_upload/view/document_upload_preparation_page.dart';
 import 'package:paperless_mobile/features/documents/view/pages/document_view.dart';
 import 'package:paperless_mobile/generated/l10n/app_localizations.dart';
-import 'package:paperless_mobile/helpers/connectivity_aware_action_wrapper.dart';
-import 'package:paperless_mobile/helpers/message_helpers.dart';
-import 'package:paperless_mobile/helpers/permission_helpers.dart';
+import 'package:paperless_mobile/core/widgets/connectivity_aware_action_wrapper.dart';
+import 'package:paperless_mobile/core/util/message_helpers.dart';
+import 'package:paperless_mobile/core/util/permission_helpers.dart';
 import 'package:paperless_mobile/routing/routes/scanner_route.dart';
 import 'package:path/path.dart' as p;
 import 'package:pdf/pdf.dart';
