+++ lib/features/documents/view/widgets/sort_documents_button.dart
@@ -5,7 +5,7 @@ import 'package:paperless_mobile/core/translation/sort_field_localization_mapper
 import 'package:paperless_mobile/features/documents/cubit/documents_cubit.dart';
 import 'package:paperless_mobile/features/documents/view/widgets/search/sort_field_selection_bottom_sheet.dart';
 import 'package:paperless_mobile/features/labels/cubit/label_cubit.dart';
-import 'package:paperless_mobile/helpers/connectivity_aware_action_wrapper.dart';
+import 'package:paperless_mobile/core/widgets/connectivity_aware_action_wrapper.dart';
 
 class SortDocumentsButton extends StatelessWidget {
   final bool enabled;
