+++ lib/features/labels/view/widgets/label_item.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:paperless_api/paperless_api.dart';
 import 'package:paperless_mobile/core/database/tables/local_user_account.dart';
-import 'package:paperless_mobile/helpers/format_helpers.dart';
+import 'package:paperless_mobile/core/util/format_helpers.dart';
 import 'package:paperless_mobile/routing/routes/labels_route.dart';
 
 class LabelItem<T extends Label> extends StatelessWidget {
