+++ lib/features/login/view/login_page.dart
@@ -9,7 +9,7 @@ import 'package:paperless_mobile/core/model/client_certificate.dart';
 import 'package:paperless_mobile/features/login/model/login_form_credentials.dart';
 import 'package:paperless_mobile/features/login/view/add_account_page.dart';
 import 'package:paperless_mobile/generated/l10n/app_localizations.dart';
-import 'package:paperless_mobile/helpers/message_helpers.dart';
+import 'package:paperless_mobile/core/util/message_helpers.dart';
 import 'package:paperless_mobile/routing/routes/login_route.dart';
 
 class LoginPage extends StatelessWidget {
